feat: article detail page (#744)
DD-348 #done 

Re-layout for article page. This includes extracting the whole article page for easier maintainability.
