+++ packages/shared/src/components/utilities.module.css
@@ -13,6 +13,14 @@
   max-width: $pageMaxWidth;
 }
 
+.newCommentContainer {
+  max-width: $pageMaxWidth;
+}
+
+.pageSidebar {
+  max-width: $pageMaxWidth;
+}
+
 :global(.non-responsive-header) ~ .pageContainer {
   @screen laptop {
     min-height: calc(100vh - $navBarHeight);
