+++ packages/shared/src/components/utilities.tsx
@@ -16,13 +16,25 @@ export const LegalNotice = classed(
 export const PageContainer = classed(
   'main',
   styles.pageContainer,
-  'relative flex flex-col w-full items-stretch px-4 z-1 tablet:px-8 tablet:self-center mx-auto',
+  'relative flex flex-col w-full items-stretch px-4 z-1 tablet:px-8 tablet:self-center',
+);
+
+export const PageSidebar = classed(
+  'aside',
+  styles.pageSidebar,
+  'laptopL:absolute right-0 px-4 tablet:px-8 order-2 laptop:pb-10 laptopL:pb-0 laptop:order-3 laptopL:order-2 laptopL:px-6 laptop:border-r laptop:border-l laptop:border-theme-divider-tertiary laptopL:border-none',
+);
+
+export const NewCommentContainer = classed(
+  'div',
+  styles.newCommentContainer,
+  'pb-20 laptop:pb-6 laptop:border-r order-3 laptop:order-2 laptopL:order-3 laptop:border-l laptop:border-theme-divider-tertiary relative flex flex-col w-full items-stretch px-4 z-1 tablet:px-8 laptopL:mr-[22.5rem]',
 );
 
 export const NoPaddingPageContainer = classed(
   'main',
   styles.pageContainer,
-  'relative flex flex-col w-full items-stretch z-1 tablet:self-center',
+  'relative flex flex-col w-full items-stretch z-1 tablet:self-center mx-auto',
 );
 
 export const ResponsivePageContainer = classed(
