+++ packages/webapp/components/posts/AuthorOnboarding.module.css
@@ -1,19 +1,3 @@
-.similarPosts {
-  @screen laptop {
-    top: 5.5rem;
-    width: 19.5rem;
-  }
-}
-
-.newComment {
-  @screen laptop {
-    left: unset;
-    right: unset;
-    bottom: unset;
-    box-shadow: none;
-  }
-}
-
 .authorOnboarding {
   & p {
     margin: 1rem 0;
@@ -31,7 +15,7 @@
   & p,
   & ol {
     color: var(--theme-label-secondary);
-    @apply typo-callout
+    @apply typo-callout;
   }
 }
 
@@ -58,12 +42,3 @@
     @apply typo-body;
   }
 }
-
-@media not all and (min-resolution:.001dpcm)
-{ @supports (-webkit-appearance:none) {
-  @media screen and (max-width: 375px){
-    .discussionBar{
-      margin-bottom: 1.5rem;
-    }
-  }
-}}
