+++ packages/webapp/components/posts/AuthorOnboarding.tsx
@@ -1,0 +1,66 @@
+import React, { ReactElement } from 'react';
+import FeatherIcon from '@dailydotdev/shared/icons/feather.svg';
+import { Button } from '@dailydotdev/shared/src/components/buttons/Button';
+import { ownershipGuide } from '@dailydotdev/shared/src/lib/constants';
+import sizeN from '@dailydotdev/shared/macros/sizeN.macro';
+import classNames from 'classnames';
+import styles from './AuthorOnboarding.module.css';
+
+interface AuthorOnboardingProps {
+  onSignUp?: () => unknown;
+}
+
+export function AuthorOnboarding({
+  onSignUp,
+}: AuthorOnboardingProps): ReactElement {
+  return (
+    <section
+      className={classNames(
+        'p-6 bg-theme-bg-secondary rounded-2xl',
+        styles.authorOnboarding,
+      )}
+    >
+      <div
+        className={classNames(
+          'grid items-center gap-x-3',
+          styles.authorOnboardingHeader,
+        )}
+        style={{ gridTemplateColumns: 'repeat(2, max-content)' }}
+      >
+        <FeatherIcon />
+        <h3>Author</h3>
+        <h2>Is this article yours?</h2>
+      </div>
+      <p>Claim ownership and get the following perks:</p>
+      <ol>
+        <li>Get notified when your articles are picked by daily.dev feed</li>
+        <li>Exclusive author badge on your comments</li>
+        <li>Analytics report for every post you wrote</li>
+        <li>Gain reputation points by earning upvotes on articles you wrote</li>
+      </ol>
+      <div
+        className="grid grid-flow-col gap-x-4 mt-6"
+        data-testid="authorOnboarding"
+        style={{
+          maxWidth: sizeN(74),
+          gridTemplateColumns: '1fr max-content',
+        }}
+      >
+        {onSignUp && (
+          <Button className="btn-primary" onClick={onSignUp}>
+            Sign up
+          </Button>
+        )}
+        <Button
+          className="btn-secondary"
+          tag="a"
+          href={ownershipGuide}
+          target="_blank"
+          rel="noopener"
+        >
+          Learn more
+        </Button>
+      </div>
+    </section>
+  );
+}
